FunctionSignatureOpts: don't remove dead owned arguments
An owned argument always has at least one consuming use.
Except if the function ends in an unreachable.
If we would remove a dead owned argument in such a function, the generated thunk would not destroy the dead owned argument.
Also, removing a dead owned argument can only be done if there is no deinit barrier between the function entry and the consuming use(s).

Fixes an ownership verification error.